requirements.txt
django-cors-headers
langchain-community
langchain-openai
psycopg2-binary
dj-database-url
langchain-core
python-dotenv
django-heroku
whitenoise
instructor
langgraph
yfinance
pydantic
requests
gunicorn
pyngrok
uvicorn
openai
Django
torch